examples: merged `error-handling` and `customize-extractor-error`
