+++ examples/customize-extractor-error/Cargo.toml
@@ -2,12 +2,16 @@
 name = "example-customize-extractor-error"
 version = "0.1.0"
 edition = "2021"
-publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-axum = { path = "../../axum" }
+axum = { path = "../../axum", features=["matched-path"] }
+axum-extra = { path = "../../axum-extra" }
+axum-macros = { path = "../../axum-macros" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-tokio = { version = "1.0", features = ["full"] }
+thiserror = "1.0"
+tokio = { version = "1.20", features = ["full"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
