--- examples/error-handling/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "error-handling"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-axum = { path = "../../axum", features=["matched-path"] }
-axum-extra = { path = "../../axum-extra" }
-axum-macros = { path = "../../axum-macros" }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-thiserror = "1.0"
-tokio = { version = "1.20", features = ["full"] }
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
